COP509-chat-app-lab/requirements.txt
openai
fastapi
uvicorn
python-dotenv
inspect-ai
inspect-evals
